Removed legacy code, moved env configurations from the starting script
